added A-frame contribution
I am especially excited about this contribution and will experiment more with A-frame because it was the first open source project I had analyzed in the beginning of this class. I had decided that it was going to be too hard of a project to contribute to based on the intimidating nature and appearance of the code on the github repository, but after taking a VR/AR workshop last week and learning more about A-frame, I think it is easier than I thought and extremely cool (up my artistic alley).